refactor structure and use futex directly for parking
